requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo[srv]==4.7.3
python-dotenv==1.0.1
certifi==2024.7.4
dnspython==2.6.1
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
gunicorn==22.0.0